+++ src/lib.rs
@@ -78,26 +78,24 @@ pub mod from_fast_noise_2;
 mod base;
 mod cellular;
 mod from_fast_noise_lite;
+mod noise;
 /// OpenSimplex2 noise functions and combinators.
 pub mod open_simplex_2;
 mod sample;
 mod seeded;
 mod tileable;
 
+pub use base::{CellDistance, CellValue, FastCellDistance, FastCellDistanceSq, FastCellValue, OpenSimplex2, OpenSimplex2s, Perlin, Simplex, Value, ValueCubic};
+pub use cellular::{CellIndex, DistanceFn, DistanceReturnType};
 #[doc(inline)]
 pub use frequency::Frequency;
+pub use noise::Noise;
 pub use noise_fn::NoiseFn;
-pub use tileable::Tileable;
-
 pub use sample::{Sample, Sample2, Sample3, Sample4, SampleWithSeed};
-pub use seeded::Seeded;
-
-pub use base::{CellDistance, CellValue, FastCellDistance, FastCellDistanceSq, FastCellValue, OpenSimplex2, OpenSimplex2s, Perlin, Simplex, Value, ValueCubic};
-
-pub use cellular::{CellIndex, DistanceFn, DistanceReturnType};
-
 #[cfg(feature = "nightly-simd")]
 pub use sample::{Sample2a, Sample3a, Sample4a};
+pub use seeded::Seeded;
+pub use tileable::Tileable;
 
 #[inline(always)]
 #[cfg(feature = "nightly-simd")]
@@ -167,47 +165,6 @@ impl core::fmt::Display for EnumFromStrError {
         f.write_str("can't convert string to enum")
     }
 }
-
-macro_rules! impl_modifier_methods {
-    () => {
-        #[inline(always)]
-        pub const fn seed(self, seed: i32) -> $crate::Seeded<Self> {
-            $crate::Seeded { noise: self, seed }
-        }
-
-        #[inline(always)]
-        pub const fn frequency(self, frequency: f32) -> $crate::Frequency<Self> {
-            $crate::Frequency { noise: self, frequency }
-        }
-
-        #[inline(always)]
-        pub const fn fbm(self, octaves: u32, gain: f32, lacunarity: f32) -> $crate::fractal::Fbm<Self> {
-            $crate::fractal::Fbm::new(self, octaves, gain, lacunarity)
-        }
-
-        #[inline(always)]
-        pub const fn ridged(self, octaves: u32, gain: f32, lacunarity: f32) -> $crate::fractal::Ridged<Self> {
-            $crate::fractal::Ridged::new(self, octaves, gain, lacunarity)
-        }
-
-        #[inline(always)]
-        pub const fn ping_pong(self, octaves: u32, gain: f32, lacunarity: f32, strength: f32) -> $crate::fractal::PingPong<Self> {
-            $crate::fractal::PingPong::new(self, octaves, gain, lacunarity, strength)
-        }
-    };
-}
-
-macro_rules! impl_modifier_methods_tileable {
-    () => {
-        #[inline(always)]
-        pub const fn tileable(self, width: f32, height: f32) -> $crate::Tileable<Self> {
-            $crate::Tileable::new(self, width, height)
-        }
-    };
-}
-
-pub(crate) use impl_modifier_methods;
-pub(crate) use impl_modifier_methods_tileable;
 
 #[cfg(test)]
 mod tests;