+++ src/seeded.rs
@@ -1,4 +1,4 @@
-use crate::{Frequency, Sample, SampleWithSeed, Tileable};
+use crate::{Noise, Sample, SampleWithSeed};
 
 /// Wraps a noise with a seed.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -7,17 +7,7 @@ pub struct Seeded<Noise> {
     pub seed: i32,
 }
 
-impl<Noise> Seeded<Noise> {
-    #[inline(always)]
-    pub const fn frequency(self, frequency: f32) -> Frequency<Self> {
-        Frequency { noise: self, frequency }
-    }
-
-    #[inline(always)]
-    pub const fn tileable(self, width: f32, height: f32) -> Tileable<Self> {
-        Tileable::new(self, width, height)
-    }
-}
+impl<N> Noise for Seeded<N> {}
 
 impl<const DIM: usize, Point, Noise> Sample<DIM, Point> for Seeded<Noise>
 where